+++ src/MasGlobal.Core/Models/DTOs/MonthlySalaryEmployeeDTO.cs
@@ -6,10 +6,10 @@ namespace MasGlobal.Core.DTOs
 {
     public class MonthlySalaryEmployeeDTO : EmployeeDTO
     {
-        public int MonthlySalary { get; set; }
+        public decimal MonthlySalary { get; set; }
 
         public MonthlySalaryEmployeeDTO(int id, string name, string contractTypeName,
-            int roleId, string roleName, string roleDescription, int monthlySalary)
+            int roleId, string roleName, string roleDescription, decimal monthlySalary)
             : base (id, name, contractTypeName, roleId, roleName, roleDescription)
         {
             MonthlySalary = monthlySalary;
